requirements.txt
cloudpickle
datasets
iopath
hydra-core
termcolor
torch
diffusers[torch]
numpy
Pillow
transformers
accelerate
wandb
torchmetrics
torch-fidelity
setuptools
bitsandbytes
PyYAML
omegaconf
torchvision
tqdm